Create ExportAllGuildsCommand.cs
Add a new command to export all Guilds. I don't really write C#, so I just combined the internal logic of GetGuildsCommand.cs and ExportGuildCommand.cs in a way that hopefully works! This addresses issues #274 and #363.

Presumably the syntax for calling this command will have to be added elsewhere, though I imagine it isn't hard.