Fixed spoilered tag and altered CSS.
The new embedded attachments tagged as spoiler will now be correctly blurred.

This method re-uses the same `div` tags, JS code and CSS for image attachments that are tagged as spoiler.

However some CSS changes were made to make them work for Firefox.
The stylistic changes that were made, unfortunately apply to other browsers too.

Changes made to the `.spoiler` class are to prevent the spoiler-overlay-div from stretching from end to end on Firefox.

`.spoiler--hidden .spoiler-image img` was renamed to `.spoiler--hidden .spoiler-image *` in order to make it work for all elements below instead of only the `img` elements.

Lastly, the `.spoiler--hidden .spoiler-image:after` class had it's `padding` changed to be 100% (the pill shaped text-button now covers the entire element), because under Firefox, this was the only workaround to prevent mouse-clicks from playing the video/audio under the blur effect.