Update Readme.md
I've added an 'and' to correct the sentence's syntax.

The sentence describes articles that DiscordChatExporter (referred to by the pronoun 'it') works with. This includes 'direct messages', 'group messages', and 'server channels'. However, the fourth article breaks out of correct syntax since it does not fall under the domain of what DiscordChatExporter works with: instead, it adds a new clause to the sentence without a valid conjunction. Thus, to fix this, we first add a valid conjunction ('and') in between the two clauses, and we add an and after the third article of the first clause as per convention. Whereas using 'and' again to refer to other rich media features is not incorrect, I've chosen to replace the conjunction in order to make the sentence flow better after accounting for the corrections.